@@ -1,6 +1,10 @@
 name: example
 description: Sample project using keyboard actions.
 
+# The following line prevents the package from being accidentally published to
+# pub.dev using `pub publish`. This is preferred for private packages.
+publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
 # followed by an optional build number separated by a +.
@@ -10,13 +14,14 @@ description: Sample project using keyboard actions.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.3.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-
-  keyboard_actions: 
+  
+  intl: any
+  keyboard_actions:
     path: ../
 
 
